requirements.txt
asgiref==3.6.0
Django==4.1.7
gunicorn==20.1.0
pyaes==1.6.1
Pyrogram==2.0.102
PySocks==1.7.1
python-dotenv==1.0.0
sqlparse==0.4.3
TgCrypto==1.2.5

